refact: extrair hooks de chat de skill para skill-chat-hooks.mjs (Fase 3 — Commit 4/4)
